{{cookiecutter.project_slug}}/bot/requirements.txt
alexandra==0.3.2
backports.shutil-get-terminal-size==1.0.0
boto3==1.3.0
botocore==1.4.27
cffi==1.9.1
coverage==4.1
cryptography==1.7.1
decorator==4.0.10
decutils==1.0
docutils==0.12
enum34==1.1.6
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.17
ipdb==0.10.1
ipython==4.2.1
ipython-genutils==0.1.0
jmespath==0.9.0
jsonte==0.8.6
lambda-uploader==1.0.0
mock==2.0.0
ordereddict==1.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.0
pickleshare==0.7.2
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
pytest==2.9.2
pytest-catchlog==1.2.2
pytest-cov==2.3.0
python-dateutil==2.5.3
sdag2==1.0.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
virtualenv==15.0.2
Werkzeug==0.11.4
